spokemon species modification moved under admin  and some major update
