+++ pokemon-spring-mvc/src/main/resources/Texts.properties
@@ -34,3 +34,10 @@ gym.list.badge.view=View my badge
 badge.view=View badge
 gym.edit.leader=Edit my gym
 gym.edit.admin=Change gym leader
+user.name=User name
+password=Password
+main.page.title.text=Project for PA165
+main.page.text=Welcome to Pokemon Leage Paricipation Manager.
+project.page=Project page.
+catch.them.all=Gotta catch 'em all!
+log.out=Log out
