+++ pokemon-spring-mvc/src/main/resources/Texts_hu.properties
@@ -34,3 +34,10 @@ gym.list.badge.view=Jelv\u00e9nyem megtekint\u00e9se
 badge.view=Jelv\u00e9ny megtekint\u00e9se
 gym.edit.leader=Termem m\u00f3dos\u00edt\u00e1sa
 gym.edit.admin=Vez\u00e9r v\u00e1ltoztat\u00e1sa
+user.name=Felhaszn\u00e1l\u00f3 n\u00e9v
+password=Jelsz\u00f3
+main.page.title.text=PA165 projekt
+main.page.text=\u00dcdv\u00f6z\u00f6lj\u00fck a Pok\u00e9mon Liga R\u00e9szv\u00e9teli Menedzserben.
+project.page=Projekt oldala.
+catch.them.all=Szerezd meg h\u00e1t mind!
+log.out=Kijelentkez\u00e9s
