+++ pokemon-spring-mvc/src/main/resources/Texts_sk.properties
@@ -1,5 +1,5 @@
 
-app.name=Mana\u017e\u00e9r \u00fa\u010dasti pokemonovej ligy
+app.name=Spr\u00e1vca \u00fa\u010dasti Pokemonovej ligy
 home=Domov
 trainer=Tr\u00e9ner
 gym=\u0160tadi\u00f3n
@@ -34,3 +34,10 @@ gym.list.badge.view=Zobrazi\u0165 m\u00f4j odznak
 badge.view=Zobrazi\u0165 odznak
 gym.edit.leader=Upravi\u0165 m\u00f4j \u0161tadi\u00f3n
 gym.edit.admin=Zmeni\u0165 ved\u00faceho
+user.name=U\u017eiv\u00e1te\u013esk\u00e9 meno
+password=Heslo
+main.page.title.text=Projekt na PA165
+main.page.text=V\u00fdtajte v Spr\u00e1vcovi \u00fa\u010dasti Pokemonovje ligy.
+project.page=Str\u00e1nka projektu.
+catch.them.all=V\u0161etky chyti\u0165 m\u00e1\u0161!
+log.out=Vyhl\u00e1si\u0165 sa
