+++ pokemon-spring-mvc/src/main/webapp/WEB-INF/jsp/pokemonSpecies/list.jsp
@@ -10,6 +10,7 @@
 <%@ taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt" %>
 <%@ taglib prefix="s" uri="http://www.springframework.org/tags" %>
 <%@ taglib prefix="form" uri="http://www.springframework.org/tags/form" %>
+<%@ taglib prefix="security" uri="http://www.springframework.org/security/tags" %>
 <fmt:setBundle basename="Texts"/>
 <fmt:message var="title" key="pokemon.species"/>
 <my:pagetemplate title="${title}">
@@ -32,27 +33,31 @@
                 <td><c:out value="${pokemonSpecies.primaryType}"/></td>
                 <td><c:out value="${pokemonSpecies.secondaryType}"/></td>
                 <td><c:out value="${pokemonSpecies.evolvesFrom}"/></td>
-                <td>
-                    <my:extraTag href="/pokemonSpecies/changeTyping/${pokemonSpecies.id}" class='btn btn-primary'>
-                        <span class="glyphicon glyphicon-edit"></span> 
-                        <fmt:message key="pokemon.species.change.typing"/>
-                    </my:extraTag>
-                </td>
-                <td>
-                    <my:extraTag href="/pokemonSpecies/changePreevolution/${pokemonSpecies.id}" class='btn btn-primary'>
-                        <span class="glyphicon glyphicon-edit"></span> 
-                        <fmt:message key="pokemon.species.change.preevolution"/>
-                    </my:extraTag>
-                </td>
+                <security:authorize access="hasRole('ADMIN')">
+                    <td>
+                        <my:extraTag href="/admin/pokemonSpecies/changeTyping/${pokemonSpecies.id}" class='btn btn-primary'>
+                            <span class="glyphicon glyphicon-edit"></span> 
+                            <fmt:message key="pokemon.species.change.typing"/>
+                        </my:extraTag>
+                    </td>
+                    <td>
+                        <my:extraTag href="/admin/pokemonSpecies/changePreevolution/${pokemonSpecies.id}" class='btn btn-primary'>
+                            <span class="glyphicon glyphicon-edit"></span> 
+                            <fmt:message key="pokemon.species.change.preevolution"/>
+                        </my:extraTag>
+                    </td>
+                </security:authorize>
             </tr>
         </c:forEach>
         </tbody>
     </table>
-    <my:extraTag href="/pokemonSpecies/new" class="btn btn-default">
-        <span class="glyphicon glyphicon-plus" aria-hidden="true"></span>
-            <fmt:message key="pokemon.species.create.new"/>
-    </my:extraTag>
+    <security:authorize access="hasRole('ADMIN')">
+        <my:extraTag href="/admin/pokemonSpecies/new" class="btn btn-primary">
+            <span class="glyphicon glyphicon-plus" aria-hidden="true"></span>
+                <fmt:message key="pokemon.species.create.new"/>
+        </my:extraTag>
+    </security:authorize>
         
-</jsp:attribute> 
+</jsp:attribute>
 </my:pagetemplate>
